transcriber/requirements.txt
numpy==1.24.3
soundfile
aiohttp==3.8.5
faster_whisper==1.1.0
google-cloud-storage==2.11.0
openai>=1.12.0
python-dotenv==1.0.1
langchain>=0.1.0
langchain-community>=0.0.16
langchain-openai>=0.0.5
chromadb>=0.6.0
tiktoken>=0.5.2
